solved

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,7 +6,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t n = 0;
+	int n = 0;
 
 	for (n = 0; h; n++)
 	{
